requirements.txt
python-dotenv
pycoingecko
flask
flask_apscheduler
sqlite3


